backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
tensorflow-cpu
nsepython